Cleaned up Users models
